feat: Added database connectivity
Created schemas for users and posts, as well as database connectivity for the cutestpawDB
Only one database connection is allowed at a time. It's currently set up for the demo DB. Change this in server.js.
Trying to use cutestpawDB anyway probably won't throw an error but nothing will be sent to it
resolves: #7