feat: Image uploading
Posts now support image uploading! I included this code in the front end as I found it much easier than the backend.
Upon calling fileUpload, the image in the file upload box will be posted to cloudinary. This is an async call to ensure the upload completes before continuing. fileUpload returns the URL of the image which is then sent to the backend.
Modify this code as needed when adding support for up to 3 images.
Additionally modified contentFeedAPI to comment out a log call, and fixed a bug in the schema