feat: visiting other profiles
Other profiles are now viewable and live at /users/username. If you visit your own profile, you will also see a logout and placeholder edit profile button. If you visit a profile of a nonexistent user, you will be redirected away.
Because hooks can't be used in class components, I had to include a wrapper function in getUsername to access useParams() and get the username of the profile trying to be accessed.
I also included a component for restricting a page to logged in users as part of the routing changes, though this won't be used on profiles.
This functionality is only present in userProfile.js. Once profile customization is complete, I will merge the functionalities together.
Fully resolves my parts of #21