feat: Uploading Images
When uploading a file, it will only accept up to three images and the
file type must be 'image'. Included error message when there is more
than three images as well as when there is no file or text in the post.
If there is more than one image the images will be displayed in a
carousel

Resolves: Issue #19