feat: implement editing profiles
Committing in lieu of Kanza who is having issues
Users may now edit their name, bio, and website. On the edit profile page, users are shown a preview of their edits as they type. If a user hasn't set an element, it won't appear.
Also merged the profile pages together, added the edit profile page to the routes, and modified the backend to work with this frontend