feat: Implement searching
Searching has been implemented. Users and posts can be sorted based on a simple text query. Posts must match the query in their postedBy (author) content, or tags. Users must match the query in their username or name.
Additionally, there is an option to return results that exactly match the search query, or include the query.
Finally, posts can be sorted by recency or popularity (most likes) in asending or decending order. Users can not be sorted.
----Other changes----
Post schema was updated to add tags to facilitate in testing. One post in the database was also tagged
Search page was added to the navbar
The profile block and user profile page were updated to add an aditional parameter, align, to specify where to float the block; left, center, or right, This was necessary so when searching users, they would display properly.