feat: Implement explore page
Pair programed with Fatema. The explore page is now live. Users can enter a tag or set of tags and be shown all images from all posts with that tag. If no images can be shown, a message will be displayed
Also added styling, added the page to routing as /explore, added the page to the navbar, and finalized backend API
resolves #40