fix: End of sprint 3 cleanup
Improved the navbar and added associated styling. The navbar now looks nicer and has a logged out state. Added it to login and signup as well
Search now does not allow withespace searches
The content feed now displays 50 posts as a comprimise between speed and user friendliness. Search does not have a limit. A page system is probably a good idea
Tested the project extensively and fixed bugs
Removed a lot of console.logs
Added some necessary libraries for pfp and banner images to package.json, as well as navbar libraries that may be useful later
resolves #55