feat: Implement account deletion
Users are now able to delete their accounts. On providing a correct password and confirming the following dialog, all of the user's presence on cutest paw will be removed, including posts, likes, comments, notifications, their name on following and follower lists, and finally the account itself.
An account settings page was added to hold the delete user component and the other 3 account settings components. It was also added to the navbar and routing.
The User API contains the deleteUser route. Finally, Packages were update to include yesnodialog, which may be a useful library, although I decided against using it.
Fully resolves #66