feat: Implemented notifications
Notifications are now implemented! Upon liking or commenting on a post or following another user, that user will be sent a notification. Notifications can be viewed at /notifications.
A user's 25 most recent notifications will appear newest first, and notifications that have appeared since the last time the user checked this page will be highlighted. If a user has no notifications, a placeholder will be displayed instead.
All notifications include a timestamp, the profile picture of the user that generated them, and a link to that user's profile. If a notification relates to a post, a link to that post will be present as well
fully resolves #59
Changes:
-Added a new field: notifCheckedTime to users to store when they last accessed the notification page
-Added notificationsAPI.js to handle getting notifications and updating when the user last accessed the notification page and added this new API to server.js
-Implemented generating notifications in postAPI (liking/commenting) and profile API (following)
-Implemented the notifications page and styling and added it to the navbar and routing
-Added placeholder notifications. notifPlaceholder isn't currently used, but could be in the future if pages are added to notifications
-Finally, fixed the bug where the profile page wouldn't refresh upon switching profiles via react-router. Passing an unused key prop to the profile page is enough to force it to re-mount.