feat: Implemented the post pages
Added the functionality to click on a post and take you to its page. Post are live at "/post/:postId". Edited the logged out states for the like and comment components, so that when logged out you can no longer add a comment or like the post. Disabled redirecting when click on components within the post component.
Styling for the post page and the post component. Also deleted left over demo files. Refactored components using post css to a new class.
Resolved: #62