feat: Final touches
Implemented some final small adjustments for use in the video and final demonstration. Committed on 4/24. This is NOT part of sprint 4 and I won't have it merged into main or deployed until that is finished being graded
Added the search page to pages accessible when logged out, fixed issues displaying a user's website when they have none, and added a link to the users profile page to their profile block